+++ src/classes/event/OnlineEvent.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 public class OnlineEvent extends Event {
     private final Location LOCATION;
 
-    public OnlineEvent(int ID, String title, LocalDateTime start, double length, Unit unit, String[] participants, Location location) {
+    public OnlineEvent(int ID, String title, LocalDateTime start, long length, Unit unit, String[] participants, Location location) {
         super(ID, title, start, length, unit, participants);
         this.LOCATION = location;
     }
